Added new namespaces, removed commented out code Removed some junk from Program.cs, added StarBot.DiscordInterop namespace for Discord facing interop specific methods.
